requirements.txt
Flask
Pillow
numpy
pypotrace
google-generativeai